+++ BanchoMultiplayerBot.Osu/Data/ScorePerformanceInfo.cs
@@ -5,8 +5,8 @@ namespace BanchoMultiplayerBot.Osu.Data;
 public class ScorePerformanceInfo
 {
     [JsonPropertyName("pp")]
-    public int PerformancePoints { get; init; }
+    public float PerformancePoints { get; init; }
     
     [JsonPropertyName("pp_max")]
-    public int MaximumPerformancePoints { get; init; }
+    public float MaximumPerformancePoints { get; init; }
 }
